自动安装lodash的invoke，build debug时不自动运行dev了，build prepare时不自动复制签名配置文件了

+++ template/build/script.build.js
@@ -16,6 +16,12 @@ var
   webpackConfig = require('./webpack.prod.conf'),
   targetPath = webpackConfig.output.path
 
+if (env.android && !shell.test('-f', path.resolve(__dirname, `../cordova/platforms/android/release-signing.properties`))) {
+  console.log('Please modify ' + 'cordova/release-signing.sample'.yellow
+  + ', then copy to ' + 'cordova/platforms/android/release-signing.properties'.yellow + ' for apk sign.')
+  process.exit()
+}
+
 shell.rm('-rf', targetPath)
 console.log(' Cleaned build artifacts.\n')
